fixed issue that caused more output to be generated with increasing number of nodes; issue with particles MPI datatype missing fields. plus comment and printed output improvements

+++ src/util.h
@@ -62,8 +62,8 @@ struct Buffers_write {
     vector<float> phi;
     
     // Buffers to fill with MPI file writing offset values
-    vector<size_t> np_count; // length of output data vecotrs for each rank
-    vector<size_t> np_offset; // cumulative sum of np_count
+    vector<int> np_count; // length of output data vecotrs for each rank
+    vector<int> np_offset; // cumulative sum of np_count
 };
 
 struct particle {
